+++ MCServerManager/ViewModels/VersionCardViewModel.cs
@@ -1,3 +1,5 @@
+using System.Threading.Tasks;
+using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using MCServerManager.Models;
 using MCServerManager.Services;
@@ -10,9 +12,32 @@ public partial class VersionItemViewModel(VersionBase version, IVersionManagerSe
 
     public VersionBase Version { get; } = version;
 
+    [ObservableProperty]
+    public partial bool InstallButtonEnabled { get; set; } = true;
+    [ObservableProperty]
+    public partial string InstallButtonText { get; set; } = "Install";
+
     [RelayCommand]
-    public void InstallVersion(string versionId)
+    public async Task InstallVersion(string versionId)
     {
-        _versionManagerService.DownloadVersion(versionId);
+        InstallButtonText = "Installing...";
+        InstallButtonEnabled = false;
+
+        bool downloadable = await _versionManagerService.DownloadVersionBinary(versionId);
+
+        if (!downloadable)
+        {
+            InstallButtonEnabled = false;
+            InstallButtonText = "Not available...";
+            return;
+        }
+
+        InstallButtonText = "Installed";
+        InstallButtonEnabled = false;
+
+        await Task.Delay(2000);
+
+        InstallButtonText = "Install";
+        InstallButtonEnabled = true;
     }
 }